+++ src/plugins/quests/cooks-assistant-quest.ts
@@ -1,8 +1,9 @@
 import { npcAction } from '@server/world/action/npc-action';
 import { dialogue, DialogueTree, Emote, execute, goto } from '@server/world/actor/dialogue';
 import { itemIds } from '@server/world/config/item-ids';
-import { PlayerQuest, Quest, QuestJournalHandler } from '@server/config/quest-config';
+import { PlayerQuest, QuestJournalHandler } from '@server/config/quest-config';
 import { Player } from '@server/world/actor/player/player';
+import { Quest } from '@server/world/actor/player/quest';
 
 
 const journalHandler: QuestJournalHandler = {
